backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==4.6.1
dnspython==2.4.2
openpyxl==3.1.2
pandas>=2.2.0
numpy>=2.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.1
# PDF QR/Barcode Extraction (Required)
PyMuPDF>=1.24.0
opencv-python>=4.9.0
Pillow>=10.0.0
pyzbar==0.1.9

# OCR Support (Optional - only needed if barcode extraction fails)
# Requires Tesseract OCR installed on system: https://github.com/tesseract-ocr/tesseract
# pytesseract==0.3.10
